+++ lib/alphabet_index_listview.dart
@@ -1,6 +1,7 @@
 library alphabet_index_listview;
 
 export 'src/anchor_scroller/anchor_scroll_controller.dart';
+export 'src/alphabet_index_header_controller.dart';
 export 'src/alphabet_stick_header_sliverview.dart';
 export 'src/alphabet_stick_header_listview.dart';
 export 'src/alphabet_index_tip_bar.dart';