+++ lib/src/alphabet_stick_header_sliverview.dart
@@ -1,145 +1,14 @@
-import 'anchor_scroller/anchor_scroll_controller.dart';
+import '../alphabet_index_listview.dart';
 import 'anchor_scroller/anchor_scroll_wrapper.dart';
 import 'alphabet_stick_header_stick.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'alphabet_index_base.dart';
-import 'alphabet_index_tool.dart';
-import 'dart:async';
-import 'dart:math';
-
-///group list view controller
-class AlphabetHeaderSliverViewController<T> {
-  ///anchor scroll controller controller
-  final AnchorScrollController _listViewController;
-
-  ///get scroll controller
-  AnchorScrollController get listviewController => _listViewController;
-
-  ///index provider
-  AlphabetHeaderProviderInterface? _headerProvider;
-
-  ///prefer group widget height
-  final double? _preferGroupHeight;
-
-  ///prefer child widget height
-  final double? _preferChildHeight;
-
-  ///scroll speed
-  final Duration _indexedScrollDuration;
-  final Curve _indexedScrollCurve;
-
-  ///create list view controller
-  AlphabetHeaderSliverViewController({
-    double? preferGroupHeight,
-    double? preferChildHeight,
-    Duration? indexedScrollDuration,
-    Curve? indexedScrollCurve,
-    required AnchorScrollController listViewController,
-  })  : _preferGroupHeight = preferGroupHeight,
-        _preferChildHeight = preferChildHeight,
-        _indexedScrollDuration =
-            (preferGroupHeight != null && preferChildHeight != null)
-                ? Duration.zero
-                : (indexedScrollDuration ?? const Duration(milliseconds: 50)),
-        _indexedScrollCurve = indexedScrollCurve ?? Curves.linear,
-        _listViewController = listViewController;
-
-  ///scroll to group
-  Future scrollToGroup(int groupIndex) async {
-    if (_headerProvider == null) {
-      return;
-    }
-
-    ///get group index
-    int index = _headerProvider!.provideIndex(groupIndex);
-
-    ///if group height prefer set
-    if (_preferGroupHeight != null &&
-        _preferGroupHeight != 0 &&
-        _preferChildHeight != null &&
-        _preferChildHeight != 0) {
-      ///get group index
-      double maxHeight =
-          _headerProvider!.provideIndexTotalGroup() * _preferGroupHeight! +
-              _headerProvider!.provideIndexTotalChild() * _preferChildHeight! +
-              _headerProvider!.provideHeightHeaderView() +
-              _headerProvider!.provideHeightFooterView() +
-              _headerProvider!.provideHeightTopPadding() +
-              _headerProvider!.provideHeightBottomPadding() -
-              _headerProvider!.provideHeightTotalList();
-      double height = groupIndex * _preferGroupHeight! +
-          (index - groupIndex) * _preferChildHeight! +
-          _headerProvider!.provideHeightHeaderView() +
-          _headerProvider!.provideHeightTopPadding();
-      listviewController.jumpTo(min(height, max(maxHeight, 0)));
-    }
-
-    ///if group height prefer not set
-    else {
-      await listviewController.scrollToIndex(
-        index: index,
-        duration: _indexedScrollDuration,
-        curve: _indexedScrollCurve,
-      );
-    }
-  }
-
-  ///scroll to child
-  Future scrollToChild(
-    int groupIndex,
-    int childIndex, {
-    Duration duration = Duration.zero,
-    Curve curve = Curves.linear,
-  }) async {
-    ///childIndex == 0 ,just scroll to group
-    if (childIndex == 0) {
-      return scrollToGroup(groupIndex);
-    }
-
-    ///get index
-    int index = _headerProvider!.provideIndex(groupIndex, child: childIndex);
-
-    ///if group height prefer set
-    if (_preferGroupHeight != null &&
-        _preferGroupHeight != 0 &&
-        _preferChildHeight != null &&
-        _preferChildHeight != 0 &&
-        duration == Duration.zero) {
-      ///get total index
-      double maxHeight =
-          _headerProvider!.provideIndexTotalGroup() * _preferGroupHeight! +
-              _headerProvider!.provideIndexTotalChild() * _preferChildHeight! -
-              _headerProvider!.provideHeightTotalList() +
-              _headerProvider!.provideHeightHeaderView() +
-              _headerProvider!.provideHeightFooterView() +
-              _headerProvider!.provideHeightTopPadding() +
-              _headerProvider!.provideHeightBottomPadding();
-      double height = groupIndex * _preferGroupHeight! +
-          (index - groupIndex) * _preferChildHeight! +
-          _headerProvider!.provideHeightHeaderView() +
-          _headerProvider!.provideHeightTopPadding();
-      listviewController.jumpTo(min(height, max(maxHeight, 0)));
-    }
-
-    ///if group height prefer not set
-    else {
-      double anchorOffset = _headerProvider!.provideHeightGroup(groupIndex);
-      await listviewController.scrollToIndex(
-        index: index,
-        duration: duration,
-        curve: curve,
-        anchorOffset: anchorOffset,
-      );
-    }
-  }
-}
 
 ///group list view
 class AlphabetHeaderSliverView<T> extends StatefulWidget {
   //controller
-  final AlphabetHeaderSliverViewController<T> controller;
+  final AlphabetHeaderViewController<T> controller;
 
   //group builder
   final AlphabetIndexGroupBuilder groupBuilder;
@@ -211,10 +80,10 @@ class _AlphabetHeaderSliverViewState<T>
       "alphabet_index_list_view_stick_header_index_prefix";
 
   ///scroll key
-  final GlobalKey _scrollKey = GlobalKey();
+  GlobalKey _scrollKey = GlobalKey();
 
   ///group key
-  final GlobalKey _groupKey = GlobalKey();
+  GlobalKey _groupKey = GlobalKey();
 
   ///header key
   final GlobalKey _headerKey = GlobalKey();
@@ -261,19 +130,6 @@ class _AlphabetHeaderSliverViewState<T>
         return AlphabetIndexTool.getItemTotalChildCount(widget.dataList);
       },
 
-      ///provide group height
-      provideHeightGroupFunc: (group) {
-        GroupPosition? groupPosition = _groupPositionList[group];
-        if (groupPosition != null) {
-          return groupPosition.endPosition - groupPosition.startPosition;
-        }
-        GroupPosition? firstOne = _groupPositionList[0];
-        if (firstOne != null) {
-          return firstOne.endPosition - firstOne.startPosition;
-        }
-        return _groupKey.currentContext?.size?.height ?? 0;
-      },
-
       ///provide total list height
       provideHeightTotalListFunc: () {
         return _scrollKey.currentContext?.size?.height ?? 0;
@@ -297,7 +153,7 @@ class _AlphabetHeaderSliverViewState<T>
         return widget.padding?.bottom ?? 0;
       },
     );
-    widget.controller._headerProvider = _headerProvider;
+    widget.controller.headerProvider = _headerProvider;
 
     ///update frame and calculate all groups position if need
     _frameUpdateListener = () {
@@ -326,8 +182,16 @@ class _AlphabetHeaderSliverViewState<T>
   void didUpdateWidget(AlphabetHeaderSliverView<T> oldWidget) {
     ///remove former listener and add current
     if (oldWidget.controller != widget.controller) {
-      oldWidget.controller._headerProvider = null;
-      widget.controller._headerProvider = _headerProvider;
+      oldWidget.controller.headerProvider = null;
+      widget.controller.headerProvider = _headerProvider;
+    }
+    if (widget.stickHeader) {
+      _scrollKey = GlobalKey();
+      _groupKey = GlobalKey();
+      widget.controller.listViewController = AnchorScrollController(
+        initialScrollOffset:
+            oldWidget.controller.listViewController.position.pixels,
+      );
     }
     WidgetsBinding.instance.addPostFrameCallback((_) {
       _groupPositionList.clear();
@@ -341,7 +205,7 @@ class _AlphabetHeaderSliverViewState<T>
   @override
   void dispose() {
     UpdateFrameTool.instance.removeFrameListener(_frameUpdateListener);
-    widget.controller._headerProvider = null;
+    widget.controller.headerProvider = null;
     _headerController.dispose();
     super.dispose();
   }
@@ -375,7 +239,7 @@ class _AlphabetHeaderSliverViewState<T>
       },
       child: CustomScrollView(
         key: _scrollKey,
-        controller: widget.controller._listViewController,
+        controller: widget.controller.listViewController,
         physics: widget.physics,
         cacheExtent: widget.cacheExtent,
         dragStartBehavior: widget.dragStartBehavior,
@@ -470,6 +334,7 @@ class _AlphabetHeaderSliverViewState<T>
         AlphabetIndexTool.getItemIndexFromGroupPos(widget.dataList, index);
     AnchorItemWrapperState? data =
         widget.controller.listviewController.itemMap[groupIndex];
+
     RenderBox? itemBox = data?.context.findRenderObject() as RenderBox?;
     Offset? offset = itemBox?.localToGlobal(const Offset(0.0, 0.0));
     if (offset != null && itemBox != null) {
@@ -488,13 +353,13 @@ class _AlphabetHeaderSliverViewState<T>
 
   ///get the group item prefer offset
   Rect? _getGroupItemPreferRect(int index) {
-    if (widget.controller._preferGroupHeight != null &&
-        widget.controller._preferChildHeight != null) {
+    if (widget.controller.preferGroupHeight != null &&
+        widget.controller.preferChildHeight != null) {
       ///calculate prefer height offset
-      double top = index * widget.controller._preferGroupHeight! +
+      double top = index * widget.controller.preferGroupHeight! +
           (AlphabetIndexTool.getItemIndexFromGroupPos(widget.dataList, index) -
                   index) *
-              widget.controller._preferChildHeight!;
+              widget.controller.preferChildHeight!;
 
       ///offset
       double offset = (widget.padding?.top ?? 0);
@@ -503,7 +368,7 @@ class _AlphabetHeaderSliverViewState<T>
         0,
         top + offset,
         MediaQuery.of(context).size.width,
-        widget.controller._preferGroupHeight!,
+        widget.controller.preferGroupHeight!,
       );
     }
     return null;
@@ -531,8 +396,8 @@ class _AlphabetHeaderSliverViewState<T>
     }
 
     //if we set the prefer child height and prefer group height, just calculate all offsets
-    if (widget.controller._preferChildHeight != null &&
-        widget.controller._preferGroupHeight != null) {
+    if (widget.controller.preferChildHeight != null &&
+        widget.controller.preferGroupHeight != null) {
       Map<int, GroupPosition> preferMap = {};
       for (int s = 0; s < widget.dataList.length; s++) {
         Rect? itemGroupRect = _getGroupItemPreferRect(s);