+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 1.0.32
+## 1.0.33
 
-1. Fix a bug that may jump to wrong index.
+1. change anchor scroll.
+2. force height setting for jump to the correct position.
+3. add curve and duration for smooth scroll.